+++ tests/test.create_chunks.R
@@ -13,23 +13,7 @@ if (file.exists("test.create_chunks.Rmd"))
 rmarkdown::draft("test.create_chunks.Rmd","github_document","rmarkdown",
                  edit=FALSE)
 
-# Create some binary data
-x = data.frame(a=1:10,b=(1:10)^2)
-saveRDS(x,"test.create_chunks.RDS")
-
-# Push chunks into Rmarkdown document
-# Insert in reverse order to not have to figure out line number
-txt = create_chunk("x\nplot(b~a,data=x)",chunk_type="r") %>%
-  insert_chunk(11,rmd.file="test.create_chunks.Rmd")
-txt = data_encode("test.create_chunks.RDS","base64") %>%
-  create_chunk(output.var="x",format="binary",loader.function=readRDS) %>%
-  insert_chunk(11,txt)
-txt = create_chunk("library(knitrdata)",chunk_type="r") %>%
-  insert_chunk(11,txt)
-
-writeLines(txt,"test.create_chunks.Rmd")
-
-# List all chunks in documents
+# List all chunks in document
 chunklst = list_rmd_chunks(file="test.create_chunks.Rmd")
 chunklst
 
@@ -39,6 +23,35 @@ txt = c(xx$pre_chunk,xx$post_chunk)
 writeLines(txt,"test.create_chunks.Rmd")
 
 # List chunks again
+chunklst = list_rmd_chunks(file="test.create_chunks.Rmd")
+chunklst
+
+# Remove all but setup chunk
+remove_chunks(file="test.create_chunks.Rmd",
+              chunk_labels = 2:nrow(chunklst),
+              output.file="test.create_chunks.Rmd")
+
+# List all chunks again
+chunklst = list_rmd_chunks(file="test.create_chunks.Rmd")
+chunklst
+
+# Create some binary data
+x = data.frame(a=1:10,b=(1:10)^2)
+saveRDS(x,"test.create_chunks.RDS")
+
+# Push chunks into Rmarkdown document
+# Insert in reverse order to not have to figure out line number
+txt = create_chunk(chunk_label="plot",c("x","plot(b~a,data=x)"),chunk_type="r") %>%
+  insert_chunk(11,rmd.file="test.create_chunks.Rmd")
+txt = data_encode("test.create_chunks.RDS","base64") %>%
+  create_chunk(chunk_label="thedata",output.var="x",format="binary",loader.function=readRDS) %>%
+  insert_chunk(11,txt)
+txt = create_chunk(chunk_label="loadknitrdata","library(knitrdata)",chunk_type="r") %>%
+  insert_chunk(11,txt)
+
+writeLines(txt,"test.create_chunks.Rmd")
+
+# List all chunks again
 chunklst = list_rmd_chunks(file="test.create_chunks.Rmd")
 chunklst